Support building on OS X
Do not use pipe2 as it is unavailable on OS X and given that
pipe2 was used without flags set to 0 there should be no
difference compared to using pipe.

Replace XPM icon that was using deprecated platform-specific icon
setting code with a PNG version of the icon instead.

Drop x.H include, not needed

ld -b deprecated with recent clang, replace with strings in the
code.Not as convenient but to make portability a non-issue just
embed the keys and license as strings in the code.