Support building with CMake
Preparing for OS X support with a build system that is available in
homebrew and helps with abstracting platform differences.
Correct link libararies are missing and missing resource compilation.

libgps not supported at all on Windows (if such a port is to be
made in the future) and unavailable in homebrew. Allow compilation
without it as florb is very usable without it.